Refactor the Project Structure.
CSVReaderUtil: Load the CSV file and map it to UserBean

UserBean: Bean Class for CSV mapping and will be used by DataProviders

moved the testData from src/main/resources to src/test/resources.
