+++ src/test/java/com/api/utils/CSVReaderUtil.java
@@ -1,24 +1,23 @@
-package com.demo.csv;
+package com.api.utils;
 
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.List;
 
+import com.dataproviders.api.bean.UserBean;
 import com.opencsv.CSVReader;
 import com.opencsv.bean.CsvToBean;
 import com.opencsv.bean.CsvToBeanBuilder;
-import com.opencsv.exceptions.CsvException;
 
-public class ReadCSVFile_MapToPojo {
-
-	public static void main(String[] args) {
+public class CSVReaderUtil {
+	
+	private CSVReaderUtil() {
 		
-		InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream("testData/LoginCreds.csv");
+	}
+	
+public static void loadCSV(String pathOfCSVFile) {
 		
+		InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream(pathOfCSVFile);
 		InputStreamReader isr = new InputStreamReader(is);
 		CSVReader csvReader = new CSVReader(isr);
